backend/requirements.txt
# Core Django and Server
Django==5.2.7
dj-database-url==2.1.0
python-decouple==3.8
python-dotenv==1.0.1
whitenoise==6.6.0
sqlparse==0.5.3
asgiref==3.10.0
pytz==2025.2
tzdata==2025.2




# DRF
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
drf-nested-routers==0.95.0
Pillow==12.0.0

# MariaDB/MySQL Support
PyMySQL==1.1.0
mysqlclient==2.2.4

# Redis
django-redis==5.4.0
redis==5.0.1

# CORS
django-cors-headers==4.3.1

# Documentation
drf-spectacular==0.27.1

# Filters
django-filter==23.5

# Google Auth
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
google-api-python-client==2.118.0
requests==2.31.0

# Ollama Cloud API
ollama==0.3.3
pydantic==2.10.3

# Celery
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.1
